Fix of calc_sign_index (#85)
* Fix of calc_sign_index
To get in line with initial Lepton implementation

* Formatting

* Some remarks (by review)

+++ src/helpers.rs
@@ -78,7 +78,7 @@ pub const fn bitn(c: u16, n: u16) -> u8 {
 }
 
 #[inline(always)]
-pub fn calc_sign_index(val: i32) -> usize {
+pub fn calc_sign_index(val: i16) -> usize {
     if val == 0 {
         0
     } else {
